testing/requirements.txt
paramiko>=3.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
click>=8.0.0
requests>=2.28.0
